Added database mdf file to source control Note: had to edit gitignore file to get this checked in.  Not sure this file is part of the solution.  Might have to do this on other machines I use.
The following link suggests not adding the database file to source control, but if I don't and send this project to someone else then they will need to run migrations to create the schema.  Whilst this should probably be the correct approach and the whole point of each migration being stored as part of the project, I don't want to complicate any instructions to someone else to get the project to work.  I want it to work out-of-the-box in this instance.
http://stackoverflow.com/questions/2738067/what-asp-net-mvc-project-files-should-i-not-add-to-subversion